requirements.txt
python-telegram-bot==20.3
beautifulsoup4==4.12.3
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
schedule==1.2.1
lxml==5.2.1